docs: actualizar README con estructura del proyecto, testing y recursos chore: configurar Jest para pruebas unitarias con React Testing Library refactor: organizar estructura del proyecto para mayor consistencia
