+++ env.local.example
@@ -1,1 +1,7 @@
+# Api
 NEXT_PUBLIC_API_URL=your_api_here
+
+# Application Settings
+NEXT_PUBLIC_APP_NAME=your_app_name_here
+NEXT_PUBLIC_ITEMS_PER_PAGE=items_per_page
+NEXT_PUBLIC_MAX_RECENT_CHARACTERS=recent_charaters
