--- jest.config.js
@@ -1,26 +1,0 @@
-/** @type {import('jest').Config} */
-module.exports = {
-  preset: 'ts-jest',
-  testEnvironment: 'jest-environment-jsdom',
-  collectCoverageFrom: [
-    'src/**/*.{ts,tsx}', 
-    '!src/**/*.styles.ts', 
-    '!src/**/*.types.ts', 
-    '!src/**/*.constants.ts',
-    '!src/**/*.selector.ts',
-    '!src/**/*.selectors.ts'
-  ],
-  moduleNameMapper: {
-    '^@/(.*)$': '<rootDir>/$1',
-    '\\.(css|less|scss|sass)$': 'identity-obj-proxy',
-    '\\.(png|jpg|jpeg|gif|svg)$': '<rootDir>/__mocks__/fileMock.js',
-  },
-  setupFilesAfterEnv: ['<rootDir>/jest.setup.ts'],
-  transform: {
-    '^.+\\.[tj]sx?$': 'babel-jest',
-  },
-  transformIgnorePatterns: [
-    '/node_modules/(?!(some-esm-module)/)'
-  ],
-  testPathIgnorePatterns: ['/node_modules/', '/.next/'],
-};
