+++ src/features/characters/components/Recent.tsx
@@ -4,7 +4,7 @@ import { useEffect, useState } from "react";
 import Image from "next/image";
 import Link from "next/link";
 import { Result } from '@/shared/types/domain';
-import { getRecentCharacters } from "@/utils/recentCharacters";
+import { getRecentCharacters } from "@/shared/utils/recentCharacters";
 
 export default function RecentCharacters() {
     const [recent, setRecent] = useState<Result[]>([]);
