+++ src/shared/components/Card/Card.styles.ts
