+++ src/shared/components/Card/Card.types.ts
