+++ src/shared/components/index.ts
@@ -2,3 +2,4 @@ export { default as Button } from './Button/Button';
 export { default as LinkButton } from './LinkButton/LinkButton';
 export { default as Input } from './Input/Input';
 export { default as Card, CardHeader, CardBody, CardFooter } from './Card/Card';
+export { default as Pagination } from './Pagination/Pagination';
